First pass on porting robot examples
