+++ newton/examples/__init__.py
@@ -183,6 +183,13 @@ def main():
         "cloth_hanging": "newton.examples.cloth.example_cloth_hanging",
         "cloth_style3d": "newton.examples.cloth.example_cloth_style3d",
         "mpm_granular": "newton.examples.mpm.example_mpm_granular",
+        "anymal_c_walk": "newton.examples.robot.example_anymal_c_walk",
+        "anymal_c_walk_physx": "newton.examples.robot.example_anymal_c_walk_physx_policy",
+        "anymal_d": "newton.examples.robot.example_anymal_d",
+        "cartpole": "newton.examples.robot.example_cartpole",
+        "g1": "newton.examples.robot.example_g1",
+        "humanoid": "newton.examples.robot.example_humanoid",
+        "quadruped": "newton.examples.robot.example_quadruped",
     }
 
     if len(sys.argv) < 2:
