+++ newton/examples/robot/example_humanoid.py
@@ -14,14 +14,14 @@
 # limitations under the License.
 
 ###########################################################################
-# Loads a MuJoCo model from MJCF into Newton and simulates it using the
-# MuJoCo solver.
+# Example Humanoid
+#
+# Shows how to set up a simulation of a humanoid articulation
+# from a USD stage using newton.ModelBuilder().
+#
 ###########################################################################
 
-import numpy as np
 import warp as wp
-
-wp.config.enable_backward = False
 
 import newton
 import newton.examples
@@ -29,16 +29,30 @@ import newton.utils
 
 
 class Example:
-    def __init__(self, stage_path="example_humanoid.usd", num_envs=8, use_cuda_graph=True):
+    def __init__(self, viewer, num_envs=4):
+        # setup simulation parameters first
+        self.fps = 60
+        self.frame_dt = 1.0 / self.fps
+
+        # group related attributes by prefix
+        self.sim_time = 0.0
+        self.sim_substeps = 10  # renamed from num_substeps
+        self.sim_dt = self.frame_dt / self.sim_substeps  # renamed from dt
+
+        # unpack any example specific args
         self.num_envs = num_envs
 
-        use_mujoco_cpu = False
+        # save a reference to the viewer
+        self.viewer = viewer
 
-        # set numpy random seed
-        self.seed = 123
-        self.rng = np.random.default_rng(self.seed)
-
-        start_rot = wp.quat_from_axis_angle(wp.normalize(wp.vec3(*self.rng.uniform(-1.0, 1.0, size=3))), -wp.pi * 0.5)
+        articulation_builder = newton.ModelBuilder(up_axis=newton.Axis.Z)
+        articulation_builder.default_joint_cfg = newton.ModelBuilder.JointDofConfig(
+            limit_ke=1.0e3, limit_kd=1.0e1, friction=1e-5
+        )
+        articulation_builder.default_shape_cfg.ke = 5.0e4
+        articulation_builder.default_shape_cfg.kd = 5.0e2
+        articulation_builder.default_shape_cfg.kf = 1.0e3
+        articulation_builder.default_shape_cfg.mu = 0.75
 
         mjcf_filename = newton.examples.get_asset("nv_humanoid.xml")
 
@@ -51,108 +65,92 @@ class Example:
             up_axis="Z",
         )
 
-        # joint initial positions
-        articulation_builder.joint_q[:7] = [0.0, 0.0, 1.5, *start_rot]
+        articulation_builder.joint_q[:3] = [0.0, 0.0, 1.3]
+        if len(articulation_builder.joint_q) > 6:
+            articulation_builder.joint_q[3:7] = [0.0, 0.0, 0.0, 1.0]
+
+        for i in range(len(articulation_builder.joint_dof_mode)):
+            articulation_builder.joint_dof_mode[i] = newton.JointMode.TARGET_POSITION
+            articulation_builder.joint_target_ke[i] = 150
+            articulation_builder.joint_target_kd[i] = 5
 
         spacing = 3.0
         sqn = int(wp.ceil(wp.sqrt(float(self.num_envs))))
 
-        builder = newton.ModelBuilder()
+        builder = newton.ModelBuilder(up_axis=newton.Axis.Z)
         for i in range(self.num_envs):
-            pos = wp.vec3((i % sqn) * spacing, (i // sqn) * spacing, 0.0)
-            articulation_builder.joint_q[7:] = self.rng.uniform(
-                -1.0, 1.0, size=(len(articulation_builder.joint_q) - 7,)
-            ).tolist()
+            pos = wp.vec3((i % sqn) * spacing, (i // sqn) * spacing, 0)
             builder.add_builder(articulation_builder, xform=wp.transform(pos, wp.quat_identity()))
+
         builder.add_ground_plane()
 
-        self.sim_time = 0.0
-        fps = 60
-        self.frame_dt = 1.0 / fps
-
-        self.sim_substeps = 10
-        self.sim_dt = self.frame_dt / self.sim_substeps
-
-        # finalize model
         self.model = builder.finalize()
+        self.solver = newton.solvers.SolverMuJoCo(self.model)
 
+        self.state_0 = self.model.state()
+        self.state_1 = self.model.state()
         self.control = self.model.control()
+        self.contacts = self.model.collide(self.state_0)
 
-        self.solver = newton.solvers.SolverMuJoCo(
-            self.model,
-            use_mujoco_cpu=use_mujoco_cpu,
-            solver="newton",
-            integrator="euler",
-            iterations=10,
-            ls_iterations=5,
-        )
+        # ensure FK evaluation (for non-MuJoCo solvers):
+        newton.eval_fk(self.model, self.state_0.joint_q, self.state_0.joint_qd, self.state_0)
 
-        self.renderer = None
-        if stage_path:
-            self.renderer = newton.viewer.RendererOpenGL(
-                path=stage_path, model=self.model, scaling=1.0, show_joints=True
-            )
+        # ensure this is called at the end of the Example constructor
+        self.viewer.set_model(self.model)
 
-        self.state_0, self.state_1 = self.model.state(), self.model.state()
+        # put graph capture into it's own function
+        self.capture()
 
-        self.use_cuda_graph = (
-            not getattr(self.solver, "use_mujoco_cpu", False) and wp.get_device().is_cuda and use_cuda_graph
-        )
-
-        if self.use_cuda_graph:
+    def capture(self):
+        self.graph = None
+        if wp.get_device().is_cuda:
             with wp.ScopedCapture() as capture:
                 self.simulate()
             self.graph = capture.graph
 
+    # simulate() performs one frame's worth of updates
     def simulate(self):
+        self.contacts = self.model.collide(self.state_0)
         for _ in range(self.sim_substeps):
-            self.solver.step(self.state_0, self.state_1, self.control, None, self.sim_dt)
+            self.state_0.clear_forces()
+
+            # apply forces to the model for picking, wind, etc
+            self.viewer.apply_forces(self.state_0)
+
+            self.solver.step(self.state_0, self.state_1, self.control, self.contacts, self.sim_dt)
+
+            # swap states
             self.state_0, self.state_1 = self.state_1, self.state_0
 
     def step(self):
-        with wp.ScopedTimer("step", active=False):
-            if self.use_cuda_graph:
-                wp.capture_launch(self.graph)
-            else:
-                self.simulate()
+        if self.graph:
+            wp.capture_launch(self.graph)
+        else:
+            self.simulate()
+
         self.sim_time += self.frame_dt
 
     def render(self):
-        if self.renderer is None:
-            return
+        self.viewer.begin_frame(self.sim_time)
+        self.viewer.log_state(self.state_0)
+        self.viewer.log_contacts(self.contacts, self.state_0)
+        self.viewer.end_frame()
 
-        with wp.ScopedTimer("render", active=False):
-            self.renderer.begin_frame(self.sim_time)
-            self.renderer.render(self.state_0)
-            self.renderer.end_frame()
+    def test(self):
+        pass
 
 
 if __name__ == "__main__":
-    import argparse
+    # Create parser that inherits common arguments and adds example-specific ones
+    # keep example options short, don't overload user with options
+    # device, viewer type, and other options are created by default
+    parser = newton.examples.create_parser()
+    parser.add_argument("--num-envs", type=int, default=4, help="Total number of simulated environments.")
 
-    parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument("--device", type=str, default=None, help="Override the default Warp device.")
-    parser.add_argument(
-        "--stage-path",
-        type=lambda x: None if x == "None" else str(x),
-        default="example_humanoid.usd",
-        help="Path to the output USD file.",
-    )
-    parser.add_argument("--num-frames", type=int, default=12000, help="Total number of frames.")
-    parser.add_argument("--num-envs", type=int, default=9, help="Total number of simulated environments.")
-    parser.add_argument("--use-cuda-graph", default=True, action=argparse.BooleanOptionalAction)
+    # Parse arguments and initialize viewer
+    viewer, args = newton.examples.init(parser)
 
-    args = parser.parse_known_args()[0]
+    # Create example and run
+    example = Example(viewer, args.num_envs)
 
-    with wp.ScopedDevice(args.device):
-        example = Example(stage_path=args.stage_path, num_envs=args.num_envs, use_cuda_graph=args.use_cuda_graph)
-
-        for frame_idx in range(args.num_frames):
-            example.step()
-            example.render()
-
-            if example.renderer is None:
-                print(f"[{frame_idx:4d}/{args.num_frames}]")
-
-        if example.renderer:
-            example.renderer.save()
+    newton.examples.run(example)
